+++ melolib/database/migrations/2023_02_19_133647_create_song_genres_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('song_genres', function (Blueprint $table) {
             $table->id();
-            $table->integer('genre_id');
-            $table->integer('song_id');
+            $table->bigInteger('genre_id');
+            $table->bigInteger('song_id');
             $table->timestamps();
         });
     }
