+++ melolib/database/migrations/2023_02_19_133730_create_song_bands_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('song_bands', function (Blueprint $table) {
             $table->id();
-            $table->integer('song_id');
-            $table->integer('band_id');
+            $table->bigInteger('song_id');
+            $table->bigInteger('band_id');
             $table->timestamps();
         });
     }
